requirements.txt
streamlit
numpy
opencv-python
pillow
glob
pandas
